Refactor Signup to fit template
