+++ client/src/components/signup/Signup.js
@@ -1,0 +1,67 @@
+import React, { useState } from 'react';
+import LoginForm from '../login/LoginForm';
+import { useMutation } from '@apollo/client';
+import { ADD_USER } from '../../utils/mutations';
+import Auth from '../../utils/auth';
+import './style/Signup.css';
+
+const Signup = (props) => {
+    const [currentForm, setCurrentForm] = useState('login');
+    const [addUser, { error }] = useMutation(ADD_USER);
+
+    const toggleForm = (formName) => {
+        setCurrentForm(formName);
+    }
+    {
+        currentForm === 'login' ? <LoginForm onFormSwitch={toggleForm} /> : <Signup onFormSwitch={toggleForm} />
+    }
+
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
+    const [username, setUsername] = useState('');
+
+    const handleSubmit = async (e) => {
+        e.preventDefault();
+        console.log("working");
+        const { data } = await addUser({
+            variables: { email, password, username }
+        });
+        Auth.login(data.addUser.token);
+        console.log(data);
+    }
+
+
+    return (
+        <>
+            <form onSubmit={handleSubmit}>
+                <div className="inner-form">
+                    <h2>Sign Up</h2>
+
+                    {/* Input field for user to enter login. */}
+                    <div className="form-input">
+                        <label>Username:</label>
+                        <input value={username} onChange={(e) => setUsername(e.target.value)} type="text" id="username" name="username" />
+                    </div>
+
+                    {/* Input field for user to enter login. */}
+                    <div className="form-input">
+                        <label>Email:</label>
+                        <input value={email} onChange={(e) => setEmail(e.target.value)} type="email"  id="email" name="email" />
+                    </div>
+
+                    {/* Input field for user to enter password. */}
+                    <div className="form-input">
+                        <label>Password:</label>
+                        <input value={password} onChange={(e) => setPassword(e.target.value)} type="password"  id="password" name="password" />
+                    </div>
+
+                    {/* Button to register a new user. */}
+                    <button name="register" type="submit">Sign Up</button>
+                    <button onClick={() => props.onFormSwitch('LoginForm')}>Already have a account? Log in here.</button>
+                </div>
+            </form>
+        </>
+    )
+};
+
+export default Signup;
