--- client/src/pages/profile/Profile.jsx
@@ -1,5 +1,0 @@
-export const Profile = () => {
-    return (
-        <>profile</>
-    )
-}
